requirements.txt
python-telegram-bot==20.3
openai==0.27.0
python-dotenv==1.0.0
